pyproject.toml
[tool.poetry]
name = "repl_python3_Bruh-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"
mcstatus = "^6.5.0"
mcrcon = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
